ssl: Remove legacy useSsl account setting
The UI for disabling client-side SSL support has been removed a long
time ago, but the setting still existed in CoreAccount and related
classes.

Since SSL is now always support in the client, remove the remnants
of the legacy setting.

Note that cores still must be able to cope with legacy clients that
have SSL disabled.