travis: Add Debian Stable and Testing to the list of CI distros
In addition to have a recent Ubuntu LTS as a baseline (we'll
switch to Bionic after the next Quassel release), supporting Debian
Stable also seems prudent. Add this distro to the CI system to
ensure that we don't break the build on Debian Stable.

Also enable Debian Testing, which should provide us with an early
warning system in case toolchain or dependency changes are about
to break Quassel.