docs: Update README.md badge, download for Actions
Update the badges in README.md to point to the new GitHub Actions CI
integration, removing the old Travis and AppVeyor badges.

Update the "Downloading" text to point to the new Windows and macOS
build artifacts provided by GH Actions.  Now we can limit shown builds
to the 'master' branch, hooray!

See: https://help.github.com/en/actions/configuring-and-managing-workflows/configuring-a-workflow#adding-a-workflow-status-badge-to-your-repository
And: https://stackoverflow.com/questions/57719605/official-badge-for-github-actions