client: Optionally ensure backlog on buffer show
Add a default-enabled option to ensure backlog is available when
showing a buffer, using the same logic as scrolling up in a buffer
whenever switching to a buffer.  No additional messages will be
fetched if the scrollbar is already visible.

This makes using Quassel without any initial backlog fetching more
convenient, as you don't need to scroll upon clicking every buffer.

Note: if all loaded messages are filtered out, nothing will appear to
be visible on the first showing, and more backlog will be fetched the
next time the buffer is shown (or if scrolled, like before).  This
happens with Quassel today already, anyways, but implementing
"fetch until visible" be a good future improvement, mirroring what
Quasseldroid does.