client: Default to AsNeeded backlog fetching
Default to the AsNeeded backlog requester for better performance with
modern (v0.13.0 or newer) Quassel cores.  Legacy cores are handled
as well.

This modifies new installs and existing configurations if the prior
default value wasn't changed.  Might be worth a mention in the release
notes given the potential performance improvement.