client: Workaround initial backlog fetch scrolling
Workaround ChatView scrolling somewhere into the middle when first
fetching enough backlog for the vertical scrollbar to become visible.

This happened before when fetching backlog by scrolling, but due to
initial backlog fetching was less noticeable.

A proper fix would be nicer, but this workaround seems to work for
now.  Feel free to tear this out later!