client: Switch infobar dialog-* icons to emblem-*
Switch the icons representing whether a feature is supported or not
from the "dialog-*" set of icons, which can be hard to distinguish at
small sizes, to the "emblem-*" set of icons, which are much more
distinguishable, and actually designed as emblems.

This means older icon sets such as Ubuntu's Humanity will no longer
have a custom icon available, but the Breeze icons aren't too far out
of place.

Oxygen does not provide these icons, so include a fallback to the
previous icon names.