settings: Add server-time to Network Features tab
Add the IRCv3 "server-time" capability to the Features UI to make it
simple to disable.

Reasoning:
IRC networks with improperly configured time may result in the wrong
time showing in Quassel.  This.. really shouldn't happen in Year 2020,
but it's come up in the freenode/#kiwiirc support channel.
(In the future, Quassel should add intelligent detection of time-sync
issues, mimicking Kiwi IRC's timedelta feature.)