cmake: Determine minimum deployment target for Qt
The minimum version of macOS supported by Qt may differ from the one
provided by the Xcode SDK in use. If started on a macOS that is older
than the minimum version indicated by Qt, the application might crash.

Qt's minimum deployment target is not available directly as a define,
variable or CMake/qmake property, but can be extracted from qmake
using a rather stupid hack. Use this to set a cache variable that
will be used in subsequent commits when generating the package's
Info.plist file.