Allow updating an ISUPPORT token.
Currently if a server makes a configuration (or entire IRCd) change
which affects any ISUPPORT (005) tokens, the entire core needs to
be restarted in order to internally reflect these changes.

With this change, upon the receipt of an ISUPPORT message, if any
tokens now differ from our internal state, they will be updated.
The best way to update these is still a reconnect of that server,
which is far better restarting the entire core.