core:ircparser: Only trim whitespace from the end of the last parameter.
QString::trimmed() removes both leading and trailing whitespace, which
is not the intention here. As QString does not have a ltrim/rtrim
function, we need to implement the equivalent here to only trim the
trailing whitespace.

As QString::trimmed() removed any whitespace for which QChar::isSpace()
returns true, I have switched the trailing whitespace check to use the
same function.

This fixes #1733.