uisupport: fix application name for .desktop shell integration
When building for KDE (cmake -DUSE_KDE=1), the KAboutData constructor
as invoked by uisupport causes the resulting application name to be
"org.kde.quassel".

At least on GNOME, this "org.kde.quassel" doesn't match the
corresponding "quasselclient.desktop" file, which means the app doesn't
get a pretty name/icon in the app launcher.

The solution is to call KAboutData::setDesktopFileName() with the
desired name.

This issue doesn't occur when building with USE_KDE=0 for two reasons:
one, because QtUiApplication already calls
QGuiApplication::setDesktopFileName() with the correct value, and two,
if desktopFileName is unset, the binary name "quasselclient" would be used
instead, which still matches "quasselclient.desktop".

An alternative workaround would be to set this flag in the
quasselclient.desktop file:

    StartupWMClass=org.kde.quassel

But I would say this is worse because the "org.kde" doesn't make sense,
since this is not a KDE project.