uisupport: Use correct application name in KAboutData
In commit 020c1634 (uisupport: fix application name for .desktop shell
integration, 2023-04-18), a call to setDesktopFileName() was added to
provide the correct .desktop file name for shell integration. However,
it also changed the application name from "quassel" to "quasselclient",
which does not match the application name Quassel sets when not built
with support for KDE, and breaks integration with KNotifications.

The reason for this is that KNotifications expects the filename of the
global notification configuration file (quassel.notifyrc) to match the
application name [1]. With "quasselclient" now set as the application name
through KAboutData, this is no longer the case, and notifications simply
do not appear.

Instead, use the previous application name, "quassel". This fixes
notifications for KDE and does not regress shell integration since we
still set the correct desktop file name.

[1] https://api.kde.org/frameworks/knotifications/html/index.html