data: Set DesktopEntry correctly in notifyrc
KNotifications requires the DesktopEntry field in notifyrc to match the
base name of the application's desktop file [1]. Monolithic installs come
with a "quassel.desktop" file, which matches what is set presently in
notifyrc. However, client installs come with "quasselclient.desktop",
and do not match.

To match up these values dynamically, introduce a notifyrc.in template
containing a placeholder which will be replaced by cmake. Since a system
can technically have both a monolithic binary and a client installed at
the same time, and we can't install two separate notifyrc files because
the application name stays the same throughout, assume the user prefers
the monolithic install when configured.

[1] https://api.kde.org/frameworks/knotifications/html/index.html