Blowfish cipher cbc/ecb key exchange compatability fix.
Previously we would use the DH1080_INIT message for ECB but
we would set the key to CBC. This lead to the other party
using ECB in their client while we were using CBC which can
lead to incorrect/failed decryption.

Key exchange now supports ECB and CBC correctly. Furthermore in the
DH key exchange, if the counterparty downgrades to ECB we allow
it for compatibility and issue a message. If the counterparty
upgrades to CBC we also allow it and inform the user.